fixed all tests

+++ todo-app/models/todo.js
@@ -1,5 +1,5 @@
 "use strict";
-const { Model } = require("sequelize");
+const { Model,Op } = require("sequelize");
 module.exports = (sequelize, DataTypes) => {
   class Todo extends Model {
     /**
@@ -8,19 +8,67 @@ module.exports = (sequelize, DataTypes) => {
      * The `models/index` file will call this method automatically.
      */
     static associate(models) {
+      Todo.belongsTo(models.User, {
+        foreignKey: "userId",
+      });
       // define association here
     }
 
-    static addTodo({ title, dueDate }) {
-      return this.create({ title: title, dueDate: dueDate, completed: false });
+    static addTodo({ title, dueDate , userId }) {
+      return this.create({ title: title, dueDate: dueDate, completed: false , userId});
     }
 
-    static async remove(id){
-      return this.destroy({ where: { id } });
+    static async remove(id,userId){
+      return this.destroy({ where: { id,userId } });
     }
 
     setCompletionStatus(value) {
       return this.update({ completed: value });
+    }
+
+    static async overdue(userId){
+      return this.findAll({
+        where: {
+          dueDate: {
+            [Op.lt]: new Date(),
+          },
+          // completed: false,
+          userId
+        },
+      });
+    }
+
+    static async dueLater(userId){
+      return this.findAll({
+        where: {
+          dueDate: {
+            [Op.gt]: new Date(),
+          },
+          userId
+          // completed: false,
+        },
+      });
+    }
+
+    static async dueToday(userId){
+      return this.findAll({
+        where: {
+          dueDate: {
+            [Op.eq]: new Date(),
+          },
+          userId
+          // completed: false,
+        },
+      });
+    }
+
+    static async completed(userId) {
+      return this.findAll({
+        where: {
+          completed: true,
+          userId
+        },
+      });
     }
 
   //   markAsCompleted() {
